[binary] Allow subcommand alias references
Adds support to the parser to allow aliases for subcommands in the
parser. At the moment, this only applies to `report`, with the 'rep'
shorthand. This is however intended for further 4.0 use via the
integration of sos-collector and soscleaner. For example, invoking 'sos
clean' and 'sos cleaner' would both be possible.

Signed-off-by: Jake Hunsaker <jhunsake@redhat.com>