[collector] Initial import of sos-collector into sos
This commit represents the first import of the sos-collector project
into sos natively for 4.0. It is not expected to be functional as of
this commit.

A minimum integration has been done at this point - namely tying
SoSCollector() as a SoSComponent() subclass, and hooking up the parser
functions.

SoSCollector will load a policy in the same way as 'sos report', and
should use the same logging methods, which will be done in a future
commit in this series.

As a divergence from the standalone sos-collector project, this
integration aims to hook in host detection with the existing Policy
model provided by sos. Additionally, we should be able to drop the
Configuration dict-subclass approach by the time this series is ready
for merge.

Related: #1988

Signed-off-by: Jake Hunsaker <jhunsake@redhat.com>