[collector] Add cluster and host loading into SoSCollector() directly
As part of moving away from the Configuration() approach, move the
loading of support host and cluster types directly into SoSCollector().

Again, `sos collect` is not expected to be functional as of this commit,
these are iterative updates to move towards that goal. Similarly, it is
still desired that by the end of this patch series, "host types" are
really just `Policy()` instances.

Related: #1988

Signed-off-by: Jake Hunsaker <jhunsake@redhat.com>