[collector] Drop Configuration class and use SoSOptions
Moves the `collect` component away from using the Configuration class
and merges relevant options into the SoSOptions framework, whereas other
options are made into either class attrs or part of a 'commons' dict
similar to what we do in `report`.

At this point, `sos collect` has basic discovery functionality, meaning
that a user can run it locally or point it to a master node via
`--master` and get a returned list of nodes that would be collected
from. However, collection is likely to fail even if the remote command
execution works.

Signed-off-by: Jake Hunsaker <jhunsake@redhat.com>