[archive|component] Move archive setup and cleanup into SoSComponent
Moves archive setup and cleanup directly into `SoSComponent()`, then has
`SoSCollector` utilize the class rather than manually creating the
archive at the end of a run. `SoSReport()` has been updated where needed
to leverage the new location of `cleanup()`.

This also adds encryption options to `SoSCollector()` to encrypt the
final archive. It may prove later than this option should also (or
instead) pass the encryption option to the sosreport runs that occur on
the hosts.

Note that this commit marks the point from which functional, complete,
runs of `sos collect` are expected to run without error and result in a
tarball of one or more collected sosreports.

Resolves: #1988

Signed-off-by: Jake Hunsaker <jhunsake@redhat.com>