[clusters|node] Fix setting node SSH key based on cluster
Previously, the ovirt (and family) cluster profile would attempt to set
a global SSH key if the ovirt-generated key was locally available by
changing the options entry for ssh_key. This no longer worked since
decoupling the shared dict, so instead create a new mechanism by which
clusters may set this.

It may prove necessary in the future to extend this functionality to
arbitrary settings, however for now I do not see that being common
place.

Signed-off-by: Jake Hunsaker <jhunsake@redhat.com>