[collector|Policies] Load a Policy rather than SoSHost within SoSNode
Rather than load a `SoSHost` instance, `SoSNode` will now load a
`Policy` that is determined by the content of `/etc/os-release` on the
remote node, as this is a standardized file for Linux-based systems and
allows for reliable identification of the host.

Each per-node policy will load the host-side package information for use
by collector in determining cluster profile and sos capabilities.

This per-node policy will not however load an `InitSystem()` or a
`ContainerRuntime()`.

Signed-off-by: Jake Hunsaker <jhunsake@redhat.com>