[sosnode] Adjust logging mechanism and levels
Changes the logging approach in `SoSNode` so that we more closely align
to the approach taken by report. That is, don't print info level
messages to console unless they are explicitly set to print to the
console - thus allowing us to get sufficient data from a run without
needing to always set verbosity higher.

Signed-off-by: Jake Hunsaker <jhunsake@redhat.com>