[sos] Do not require collect to be available for base sos runs
Not all distributions want to add a dependency on pexpect as a default
for the installation of sos, as report does not require it.

As such, if we fail to import the collector module, add placeholder
psuedo-components so that --help output (and attempts to run the command
outright) still indicate that the functionality is available in some
manner, just with an additional step from the user.

This covers two distinct failures during the collector import:

	1. The collector module is not present. This implies that the
	   module has been split into sub-packages. The user will need
	   to install the appropriate sub-package for their distro.

	2. The pexpect package is not present. The user will just need
	   to install the python3 pexpect package.

Signed-off-by: Jake Hunsaker <jhunsake@redhat.com>