[collector] Fix local collections requiring sudo
Local report generation when running as a non-root user were broken
following the transition away from the Configuration() approach. Now,
adjust to the new way we execute so that we properly detect non-root
runs, prompt for a sudo password if needed, and/or fail appropriately if
we cannot use sudo locally for generating the report.

Signed-off-by: Jake Hunsaker <jhunsake@redhat.com>