[collector] Use jbon cluster is none is determined but we have nodes
If we end up with no cluster determination, but have a manual list of
nodes provided, then attempt to continue by setting the cluster type to
JBON (just a bunch of nodes - our 'none' cluster).

This will not address situations where a cluster type cannot be
determined, and the node strings are all regexes.

Signed-off-by: Jake Hunsaker <jhunsake@redhat.com>