[sosnode] Fix conflicts between password-per-node and sudo options
Under specific circumstances, we previously could end up prompting for
SSH passwords for nodes that would then be overridden by
--password-per-node, which in turn could potentially cause issues for
use cases where sudo was used both locally and remotely where the sudo
passwords differed (as indicated by password-per-node).

Adjust prompting and setting of the sudo passwords so that we prompt the
user the correct (minimal) amount of times, and the correct passwords
are used for sudo usage (when needed).

Signed-off-by: Jake Hunsaker <jhunsake@redhat.com>