[sosnode] Fix checking sos version from container image
Since hooking into Policy(), checking the version of sos installed in a
container on a containerized host was failing. Add an explicit override
check for this to the policies that need it, rather than bending
PackageManager to also allow container based checks since all other
package checks will still need to be done against the host and not the
container.

Signed-off-by: Jake Hunsaker <jhunsake@redhat.com>