[collect] When collect is unavailable, don't fail on unknown options
When `collect` is unavailable locally, we should always print the
unavailable message in all cases. However, due to the way the parser was
configured, any options passed to `sos collect` would trigger argparse's
"Unknown argument" error output instead.

Now, capture the unknown args when we perform the first parsing to
capture the component, then for cases where `collect` is unavailable,
simply ignore the parser to always force our own error message to be
printed.

This commit marks the end of the patchset for collect integration into
sos.

Closes: #1988
Resolves: #2017

Signed-off-by: Jake Hunsaker <jhunsake@redhat.com>