[Plugin] Apply size limit to directory specs
Addresses an issue where size limiting was not being applied to whole
directory specifications, as we would only stat the directory directly
when it was returned from a glob. Now, if the given copyspec is just a
directory path, recurse into it with an explicit '*' match.

Updates the forbidden path check to account for the newly extended files
list to match against the directory-only forbidden specifications. This
may result in more log entries for deep directory structures whose
parent directory/directories is/are forbidden.

Closes: #1750
Resolves: #2035

Signed-off-by: Jake Hunsaker <jhunsake@redhat.com>