[openshift] Update plugin for OCP4.x
This patch updates the current `openshift` plugin from 2.x (now EOL) to
4.x (3.x is handled via the `origin` plugin). This has been done at the
request of the Red Hat OCP support team.

While OCP 4.x is still built on top of kubernetes, the collection
requirements differ far to much to be able to simply extend the
kubernetes plugin. Most notably, the OCP4 plugin is designed to collect
a significantly greater volume of data by default due to the complexity
of even a basic installation of OCP4 clusters. Additionally, there are
API login requirements to obtain `oc` command output even as root, which
is not the case with base kubernetes. In order to use the API, users
should supply a bearer token either via the `openshift.token` plugin
option, or by setting the `SOSOCPTOKEN` environment variable to avoid
the token being included in output collected in the archive.

By default, the plugin will capture cluster-level information under the
'cluster_resources' directory within the plugin directory, organized in
subdirs by resource names. Similarly, each default namespace (project)
that exists at deployment time will be collected from, with each
namespace having a distinct subdir under the 'namespaces' directory
within the plugin directory. This list of namespaces to collect from can
be added to or overriden with the `--add-namespaces` and
`--only-namespaces` options respectively.

During testing it was observed that over 50 of these default namespaces
exist out of the box, and as such both the setup() and collect() phases
can take several minutes each. Total runtime of just the openshift
plugin from start to finish averages around 10 minutes, and so if the
plugin collects from the API a notification is printed to indicate this
so that end users do not erroneously think sos has hung.

Signed-off-by: Jake Hunsaker <jhunsake@redhat.com>