[kvm] change check_enabled to /dev/kvm
The KVM plugin get triggered in a container
(e.g lxd) because of "/sys/module/kvm"
inheritance from the kernel host.

Not only it's a waste of sosreport time,
but running it inside a container may
unintentionnaly reveal details from its
host. Which is a undesired behaviour.

Switching to /dev/kvm, is more appropriate
and follow current standard as used by tool
such as cpu-checker (kvm-ok) for instance.

And taking benefit of this change to get rid
of the check_enabled() overwrite in favor of
using "files=" trigger.

Closes: #2062
Resolves: #2063

Signed-off-by: Eric Desrochers <eric.desrochers@canonical.com>
Signed-off-by: Jake Hunsaker <jhunsake@redhat.com>