[report] Add hook into SoSCleaner
Adds a new option, `--clean` or `--mask`, to hook a report into
`SoSCleaner`.

This is done after collection is complete, and before the archive is
built and compressed. Data is substituted in place for the report, which
means there will not be an obfuscated copy on disk. There will however
still be a mapping file produced and located in the same directory as
the final archive.

If it desired to have both an obfuscated and unobfuscated copy, it is
recommended to run `sos report` followed by a separate `sos clean`.

Related: #1987

Signed-off-by: Jake Hunsaker <jhunsake@redhat.com>