Clarify default timeout for all plugins
The default timeout for all plugins is 300, but
the global timeout is set as -1 in the option list. This is
because this timeout is unset by default. This patch
attempts to clarify the -1 exposed to the user when using
commands like 'sos report -l', via the option description
as well as a note clarifying it in the man page.

Closes: #2003

Signed-off-by: Jose Castillo <jcastillo@redhat.com>