[Plugin] Don't duplicate non-empty regex directory copyspec expansion
As part of #2035, we recurse into directory copyspecs in order to
properly enumerate all files under that path for size limiting reasons.
However, it was found that copyspecs that are regexes that match only
directories get added effectively twice - once for the top-level
directory directly and then again for the files underneathe.

Fix this by removing the top-level path from the returned set of paths
to collect from any given copyspec, if that top-level path is a
non-empty directory.

Resolves: #2089

Signed-off-by: Jake Hunsaker <jhunsake@redhat.com>