[selinux] Increase default size limit for /var/lib/selinux
Increase the default size limit for /var/lib/selinux, as the global
default of 25 MB is too small to collect both the 'active' and
'targeted' subdirs in many test scenarios. Previously this had been
copied in full due to the size limiting bug that #2035 resolved, so now
we need to properly scope the size limit for this collection.

Resolves: #2090

Signed-off-by: Jake Hunsaker <jhunsake@redhat.com>