[cleaner] Remove comment filtering, improve obfuscation logging
First, stop filtering out comments and process them the same as any
other line.

Second, improve the logging made for obfuscating files by including the
archive name and the relative path, rather than the absolute path within
the tmpdir we use. Note that this also changed the format of files to
skip to the relative path within the archive.