[report] Immediately exit on first fatal error from plugins
Updates our exit path when we encounter a fatal filesystem error
(ENOSPC, EROFS) to use `os._exit(1)` rather than our own `_exit()`
method so that we can properly exit from the ThreadPool.

This does leave the temporary directory in place on the filesystem, so
the exit message has been updated to highlight that fact to the end
user.

Closes: #2071
Resolves: #2108

Signed-off-by: Jake Hunsaker <jhunsake@redhat.com>