[collect] Pass SoSNode object for remote policy loading
Rather than read `/etc/os-release` indiscriminately and pass the entire
contents as a single string to each policy's `check()` method, now just
pass the actual `SoSNode()` object and allow the policy to determine
what to do with it, if provided.

This allows policies to mimic their local checks for remote nodes, be
that substring presence in a file, a file's presence, or anything else.

Resolves: #2111

Signed-off-by: Jake Hunsaker <jhunsake@redhat.com>