[lvm2] Fix warning due to locking_type=0 option
On RHEL 8, the option `--config="global{locking_type=0}"` is deprecated.
This option is included in the lvm_opts variable and thus produces
deprecation warnings in the outputs of pvscan, vgscan, pvs, vgs, and
lvs:

    WARNING: locking_type (0) is deprecated, using --nolocking.

This patch fixes the issue by removing "locking_type=0" from the config
string and appending "--nolocking" to lvm_opts on systems where
"--nolocking" is supported.

Related to RHBZ#1849248.
Resolves: #2127

Signed-off-by: Reid Wahl <nrwahl@protonmail.com>