[collect] Update host group loading from /etc/sos
Updates `sos collect` to read and write host group configurations to
`/etc/sos/groups.d/` if run as root, or `~/.config/sos/groups.d/` if run
as non root. Note that non-root users will be able to read host group
files in `/etc/sos` to allow for easier inter-user sharing, but non-root
users will not be able to write or update config files in that location.

The order of precedence for loading host groups is:

  1. The value of `--group` is an actual file path
  2. The group definition exists under ~/.config/sos/groups.d/
  3. The group definition exists under /etc/sos/groups.d/

Related: #2125

Signed-off-by: Jake Hunsaker <jhunsake@redhat.com>