[build] Remove sos.conf from setup.py, make config file optional
In the recent packaging update (#2160), sos.conf was added to `setup.py`
to include it in the source distribution tarball produced. That was the
wrong approach, as setuptools explicitly tries to prevent files from
being written outside of the package directory.

Instead, the correct approach is to make the config file effectively
optional, by producing a warning message when it doesn't exist, but
still continue on with the defaults rather than aborting.

Note that parsing errors and duplicate entries will still cause sos to
abort execution.

Resolves: #2174

Signed-off-by: Jake Hunsaker <jhunsake@redhat.com>