[Plugin] Add default tags to command and file collections
By default, now add metadata tags for the generated manifest for command
and file collections.

Commands will default to being tagged with the command being run. E.G.
running `uname -a` will get a tag of `uname`. Commands run via
`add_blockdev_cmd()` will also have the device name added to the tags.

File collections will be tagged iff the copyspec used in the
`add_copy_spec()` call matched a single file, in which case the tag will
be the actual filename, with `-` characters converted to `_`. Excepting
`.conf` files, file extensions are stripped. For these files, `_conf` is
added to the filename tag.

This is being used as a best-effort default behavior, and individual
commands and files may still be explicitly tagged by passing those
value(s) to `tags` in the appropriate collection methods. Any tags added
this way will be appended to the defaults, if any defaults are applied.

Signed-off-by: Jake Hunsaker <jhunsake@redhat.com>