[Plugin] Add regex matching for adding tags
For some time now, the prevailing style in sos has been to group
commands and file collections into lists, and pass those lists into
single `add_copy_spec()` or `add_cmd_output()` calls. This however
presents an issue when trying to retroactively add tags to these
commands or files individually as the `tags` argument is applied against
the whole call.

To fix this, without having to break out commands or creating confusing
for loops, add a new `add_cmd_tags()` method that takes a `dict` of
command regex keys, whose values are the tag(s) that should be added to
commands matching those regexes.

As tags are determined at collection time, and not during `setup()`,
this method may be called at any point during `setup()` regardless of
when the `add_cmd_output()` calls are made.

Signed-off-by: Jake Hunsaker <jhunsake@redhat.com>