[Plugin] Add method to tag files collected by generic copyspecs
Most of the files collected by sos are collected by 'generic' copyspecs,
such as those that target a parent directory and copy the contents
recursively. This makes it difficult to selectively add tags to files
copied from these copyspecs, without adding that tag to _every_ file
collected by it.

To alleviate this, add a new `add_file_tags()` method to `Plugin` that,
like `add_cmd_tags()` takes a dict of file name regex keys whose values
are the tag(s) to add to those files in the manifest.

Note that due to how file collection occurs, this tagging is done after
all collections are complete and any matches are added as _additional_
entries to the manifest, whose specification will be the regex used to
apply the tag.

Signed-off-by: Jake Hunsaker <jhunsake@redhat.com>