[archive] Fix exception handling and LGTM alerts
First, consolidate a catch for IOError and OSError into a single OSError
catch, as since python-3.3 IOError has been merged into OSError. This in
turn resolves a LGTM alert for an unreachable exception catch.

Second, disable the init-calls-subclass warning as it is being triggered
by part of the test suite which for our purposes is a false-positive.

Resolves: #2225

Signed-off-by: Jake Hunsaker <jhunsake@redhat.com>