[obfuscation_archive] Fix tarball skipping regex
The regex used to skip tarballs was slightly too vague, in that we were
checking for any characters following 'tar', instead of any characters
following 'tar.'. This in turn lead to the potential for skipping files
that included the original sosreport archive's name.

Make the regex more exact to no longer make these accidental matches,
and only match on true tarballs.

Closes: #2236
Resolves: #2238

Signed-off-by: Jake Hunsaker <jhunsake@redhat.com>