[rabbitmq] Add 10 second timeout to call to `maybe_stuck()`.
The `maybe_stuck()` function might run for a long time if it is
tracking down a lot of potentially stuck processes. In order to
prevent long run times of `sos report` the `cmd` timeout should be set
explicitly. As a compromise between acceptable execution times and
information gathered from the plugin, this change introduces a timeout
of 10 seconds for the `maybe_stuck()` call.

Resolves: #2239

Signed-off-by: Nicolas Bock <nicolas.bock@canonical.com>
Signed-off-by: Jake Hunsaker <jhunsake@redhat.com>