[policy] Handle additional failure conditions for FTP uploads
Adds a timeout and a timeout handler for FTP connections, rather than
letting the connection attempt continue indefinitely.

Second, adds exception handling for an edge case where the connection to
the FTP server fails, but does not generate an exception from the ftplib
module.

Additionally, correct the type-ing of the error numbers being checked so
that we actually match them.

Resolves: #2245

Signed-off-by: Jake Hunsaker <jhunsake@redhat.com>