[plugins] Use `IndependentPlugin` tag in more plugins
This commit updates 94 plugins to use `IndependentPlugin` as the tagging
class, instead of the distro-specific tags.

This tag is applied to plugins where _all_ of `RedHatPlugin`,
`DebianPlugin`, and `UbuntuPlugin` are imported and there is only a
single class definition for the plugin which uses all of those tagging
classes. `SuSEPlugin` is also accepted where it is explicitly imported,
as the SuSE policy already uses the `RedHatPlugin` tagging class as
well.

Resolves: #2256

Signed-off-by: Jake Hunsaker <jhunsake@redhat.com>