[openvswitch] fix dpdk-devbind.py command
dpdk_nic_bind.py came from DPDK project and was renamed
to dpdk-devbind.py in DPDK 16.07.

The renamed script uses a hyphen as a seperator and
not an underscore.

Checked upstream DPDK project and packaged versions
on Fedora32/RHEL8/Ubuntu20.04 and all using hyphen.

Replace underscore with hyphen in the sos command.

Signed-off-by: Kevin Traynor <ktraynor@redhat.com>
Signed-off-by: Jake Hunsaker <jhunsake@redhat.com>