[sos] Fix use of relative paths for --tmp-dir
Fixes an issue where the use of relative paths for `--tmp-dir` causes a
failure in the building of the final archive. Previously, a relative
path would cause the tarball to be produced in a nested directory under
the temp directory of the same name, which would in turn cause
compression and all further operations for the archive to fail.

Fix this by converting relative paths in the option to the absolute path
internally.

Resolves: RHBZ#1891562

Signed-off-by: Jake Hunsaker <jhunsake@redhat.com>