[report] improve ThreadPoolExecutor use and style
PR: #2294

Changing three things to conform more to the ThreadPoolExecutor
documentation (https://docs.python.org/3.6/library/concurrent.futures.html):

 * Using the ContextManager capabilities by applying a `with` statement.
   The advantage of that version is that the cleanup code will be executed even
   if an exception occurs during execution.
 * I removed it as attribute of the Report object. I only saw it being used in
   these few lines, so it's not necessary to have this variable to exist
   outside the context where it was used.
 * I give it the pluglist as expanded list.

The last point shouldn't be underestimated and might result in code breaking
for users (e.g. two parallel threads accessing the same file now). In the old
version the whole `pluglist` was treated as one item that was given to one
thread of `_collect_plugin`. In the new version *each item* of `pluglist` is
handed separately to its own thread of `_collect_plugin`. I think this was the
intention of using ThreadPoolExecutor in the first place. But if the items are
not split up by expanding the list, actually only one thread would run and not
"up to max_workers" in parallel.

Example:

```
>>> def asdf(a, *b):
...   print(str(a))
...   print(str(b))
...
>>> asdf(1,[2,3,4,5])
1
([2, 3, 4, 5],)    # <----- the whole list is given as one item in the tuple
>>> asdf(1, *[2,3,4,5])
1
(2, 3, 4, 5)      # <-----  each item is treated separately in the tuple so
                            map() can apply "1" to each item individually
```

Signed-off-by: Erik Bernoth <ebernoth@redhat.com>