[component] execute() as Abstract class
Resolves: #2295

`execute()` is a method that should be implemented by subclasses of
Component. And the python documentation recommends to also add the
method to the parent class and raise a specific error to indicate
that the parent class should not implement it.

See: https://docs.python.org/3/library/exceptions.html#NotImplementedError

Signed-off-by: Erik Bernoth <ebernoth@redhat.com>
Signed-off-by: Jake Hunsaker <jhunsake@redhat.com>