[sos] Use parse_args() instead of parse_known_args()
When sos-4.0 work first went in, the way we extracted cmdline args for
the parser was to call `parse_known_args()` in an effort to avoid
passing unknown args to subcmds that don't support those specific args.

As the parsers were built out however, this became unnecessary. The
default behavior of `parse_args()` is smart enough to filter out args
that are not supported by the specific component being run.

With this being the case, we no longer need to manually check for
unknown args being passed, and with that clear a LGTM warning.

Resolves: #2299

Signed-off-by: Jake Hunsaker <jhunsake@redhat.com>