[sos] Fix argparse error output formatting
Error message propagation from the subparsers was causing the error
message to be mangled with the usage string from the "main" parser when
an invalid argument value was passed.

After discussion with python upstream, while this is an issue with the
subparser formatting (as %(prog) is being set to the main parser's
usage), there are simple paths around this behavior so a code change
there will not be forthcoming.

Instead, override the subparser's prog to be a single line string
instead of the main parser's usage string.

Closes: #2285
Resolves: #2301

Signed-off-by: Jake Hunsaker <jhunsake@redhat.com>