[component] Add log verbosity from presets
Closes: #2289

The main problem this tries to solve was that preset verbosity was
ignored in logging.

With a simple test this could be reproduced:
sudo sh -c "source /path/to/repo/sosreport/venv/bin/activate; \
            cd /tmp/foo; sos report --preset sostestpreset; cd -"

The bug is that without a change of code there are no messages from the
plugin `host` (no lines of output start wiht "[plugin:host]").

The problem is that the logging is set in the inherited __init__() method
from Component, but the presets are only handled afterwards in the
Report's __init__().

Since it is good to have logging configured from the beginning, the
only option is to reconfigure it after the preset config is known.

The simplest method is to reinitialize the logging, although maybe not
the most efficient.

Signed-off-by: Erik Bernoth <ebernoth@redhat.com>