[Cirrus] Add testing via Cirrus CI
Adds functional CI testing via Cirrus in conjunction with GCE. Tests can
now be run on Fedora as well as Ubuntu hosts, and minor tasks can be run
in containers. This will allow us to easily expand our testing base
across more distributions provided those distributions can be run on GCP
instances.

As new releases of supported distributions are made available, the
maintainers will need to build and push updated Fedora (or more
generally, RH-family) images to the GCP project. Ubuntu has cloud images
on GCE already, so when new releases of Ubuntu are pushed we will simply
need to update .cirrus.yml to point to the new public images.

Note that at the moment testing on RHEL is not enabled, though it should
follow the same framework as the Fedora tests and should hopefully be
coming before too long.

Closes: #1885
Resolves: #2305

Signed-off-by: Jake Hunsaker <jhunsake@redhat.com>